+++ packages/relay/tests/lib/eth/eth_estimateGas.spec.ts
@@ -33,17 +33,8 @@ let ethImplOverridden: Eth;
 let eventEmitter: EventEmitter;
 describe('@ethEstimateGas Estimate Gas spec', async function () {
   this.timeout(10000);
-  const {
-    restMock,
-    web3Mock,
-    hapiServiceInstance,
-    ethImpl,
-    cacheService,
-    mirrorNodeInstance,
-    logger,
-    registry,
-    commonService,
-  } = generateEthTestEnv();
+  const { restMock, web3Mock, hapiServiceInstance, ethImpl, cacheService, mirrorNodeInstance, logger, commonService } =
+    generateEthTestEnv();
 
   const requestDetails = new RequestDetails({ requestId: 'eth_estimateGasTest', ipAddress: '0.0.0.0' });
   eventEmitter = new EventEmitter();
@@ -88,7 +79,6 @@ describe('@ethEstimateGas Estimate Gas spec', async function () {
       mirrorNodeInstance,
       logger,
       '0x12a',
-      registry,
       cacheService,
       eventEmitter,
     );
