+++ packages/relay/tests/lib/openrpc.spec.ts
@@ -129,15 +129,7 @@ describe('Open RPC Specification', function () {
     sdkClientStub = sinon.createStubInstance(SDKClient);
     sinon.stub(clientServiceInstance, 'getSDKClient').returns(sdkClientStub);
     // @ts-ignore
-    ethImpl = new EthImpl(
-      clientServiceInstance,
-      mirrorNodeInstance,
-      logger,
-      '0x12a',
-      registry,
-      cacheService,
-      eventEmitter,
-    );
+    ethImpl = new EthImpl(clientServiceInstance, mirrorNodeInstance, logger, '0x12a', cacheService, eventEmitter);
 
     // mocked data
     mock.onGet('blocks?limit=1&order=desc').reply(200, JSON.stringify({ blocks: [defaultBlock] }));
